First step in replacing grid handler by list panel
